src/presentation/controllers/login/login-controller.ts
import { Controller, HttpRequest, HttpResponse, IAuthentication } from './login-protocols'
import { badRequest, ok, serverError, unauthorized } from '@/presentation/helpers/http/http-helper'
import { IValidation } from '../../protocols/validation'

export class LoginController implements Controller {
  constructor (private readonly validation: IValidation, private readonly authentication: IAuthentication) { }

  async handle (httpRequest: HttpRequest): Promise<HttpResponse> {
    try {
      const error = this.validation.validate(httpRequest.body)
      if (error) {
        return badRequest(error)
      }
      const { email, password } = httpRequest.body
      const authenticationModel = await this.authentication.auth({ email, password })
      if (!authenticationModel) {
        return unauthorized()
      }
      return ok(authenticationModel)
    } catch (error) {
      return serverError(error)
    }
  }
}
